chore: make sure error checking only runs for debug

+++ src/common/renderer.cpp
@@ -28,7 +28,7 @@ void Renderer::draw(const VertexArray &vertexArray, const IndexBuffer &indexBuff
     vertexArray.bind();
     indexBuffer.bind();
 
-    GLCall(glDrawElements(GL_TRIANGLES, indexBuffer.getCount(), GL_UNSIGNED_INT, nullptr));
+    GLCall(glDrawElements(GL_TRIANGLES, static_cast<int>(indexBuffer.getCount()), GL_UNSIGNED_INT, nullptr));
 }
 
 void Renderer::clear() const { glClear(GL_COLOR_BUFFER_BIT); }
